# docs/swagger.yaml
basePath: /
definitions:
...
  model.Asset:
    properties:
      business_line:
        type: string
      category:
        type: string
      created_at:
        type: string
      department:
        type: string
      description:
        type: string
      domain:
        type: string
      id:
        type: integer
      importance_level:
        description: 1-高，2-中，3-低
        type: integer
      ip_address:
        type: string
      name:
        type: string
      os:
        type: string
      owner:
        allOf:
        - $ref: '#/definitions/model.User'
        description: 关联
      owner_id:
        type: integer
      port:
        type: string
      status:
        description: 1-正常，0-下线
        type: integer
      type:
        type: string
      updated_at:
        type: string
      version:
        type: string
    required:
    - name
    - type
    type: object
...
  model.AssetCreateRequest:
    properties:
      business_line:
        type: string
      category:
        type: string
      department:
        type: string
      description:
        type: string
      domain:
        type: string
      importance_level:
        enum:
        - 1
        - 2
        - 3
        type: integer
      ip_address:
        type: string
      name:
        maxLength: 100
        minLength: 1
        type: string
      os:
        type: string
      owner_id:
        type: integer
      port:
        type: string
      type:
        enum:
        - server
        - database
        - application
        - network
        type: string
      version:
        type: string
    required:
    - name
    - type
    type: object
  model.AssetImportRequest:
    properties:
      assets:
        items:
          $ref: '#/definitions/model.AssetCreateRequest'
        type: array
    required:
    - assets
    type: object
  model.AssetInfo:
    properties:
      business_line:
        type: string
      category:
        type: string
      created_at:
        type: string
      department:
        type: string
      description:
        type: string
      domain:
        type: string
      id:
        type: integer
      importance_level:
        type: integer
      ip_address:
        type: string
      name:
        type: string
      os:
        type: string
      owner:
        $ref: '#/definitions/model.UserInfo'
      owner_id:
        type: integer
      port:
        type: string
      status:
        type: integer
      type:
        type: string
      updated_at:
        type: string
      version:
        type: string
    type: object
...
  model.AssetStats:
    properties:
      by_department:
        additionalProperties:
          format: int64
          type: integer
        type: object
      by_importance:
        additionalProperties:
          format: int64
          type: integer
        type: object
      by_status:
        additionalProperties:
          format: int64
          type: integer
        type: object
      by_type:
        additionalProperties:
          format: int64
          type: integer
        type: object
      total:
        type: integer
    type: object
...
  model.AssetUpdateRequest:
    properties:
      business_line:
        type: string
      category:
        type: string
      department:
        type: string
      description:
        type: string
      domain:
        type: string
      importance_level:
        enum:
        - 1
        - 2
        - 3
        type: integer
      ip_address:
        type: string
      name:
        maxLength: 100
        minLength: 1
        type: string
      os:
        type: string
      owner_id:
        type: integer
      port:
        type: string
      status:
        enum:
        - 0
        - 1
        type: integer
      type:
        enum:
        - server
        - database
        - application
        - network
        type: string
      version:
        type: string
    type: object
...
  model.PaginationResponse:
    properties:
      data: {}
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  model.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  model.Role:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      permissions:
        items:
          type: string
        type: array
      updated_at:
        type: string
    required:
    - name
    type: object
...
  model.RoleInfo:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      permissions:
        items:
          type: string
        type: array
    type: object
...
  model.User:
    properties:
      avatar:
        type: string
      created_at:
        type: string
      department:
        type: string
      email:
        type: string
      id:
        type: integer
      last_login_at:
        type: string
      phone:
        type: string
      real_name:
        type: string
      role:
        allOf:
        - $ref: '#/definitions/model.Role'
        description: 关联
      role_id:
        type: integer
      status:
        description: 1-正常，0-禁用
        type: integer
      updated_at:
        type: string
      username:
        type: string
    required:
    - email
    - real_name
    - role_id
    - username
    type: object
...
  model.UserInfo:
    properties:
      avatar:
        type: string
      department:
        type: string
      email:
        type: string
      id:
        type: integer
      last_login_at:
        type: string
      phone:
        type: string
      real_name:
        type: string
      role:
        $ref: '#/definitions/model.RoleInfo'
      status:
        type: integer
      username:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@vulnark.com
    name: VulnArk Team
    url: https://vulnark.com
  description: VulnArk漏洞管理系统API文档
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://vulnark.com/terms
  title: VulnArk API
  version: 1.0.0
paths:
...
  /api/v1/assets:
    get:
      description: 获取资产列表（分页）
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: 搜索关键词
        in: query
        name: keyword
        type: string
      - description: 资产类型
        in: query
        name: type
        type: string
      - description: 资产分类
        in: query
        name: category
        type: string
      - description: 所属部门
        in: query
        name: department
        type: string
      - description: 重要性等级
        in: query
        name: importance_level
        type: string
      - description: 资产状态
        in: query
        name: status
        type: string
      - description: 负责人ID
        in: query
        name: owner_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.PaginationResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取资产列表
      tags:
      - 资产管理
    post:
      consumes:
      - application/json
      description: 创建新的资产
      parameters:
      - description: 创建资产请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.AssetCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Asset'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: 创建资产
      tags:
      - 资产管理
  /api/v1/assets/{id}:
    delete:
      description: 删除资产
      parameters:
      - description: 资产ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: 删除资产
      tags:
      - 资产管理
    get:
      description: 根据ID获取资产详情
      parameters:
      - description: 资产ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.AssetInfo'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取资产详情
      tags:
      - 资产管理
    put:
      consumes:
      - application/json
      description: 更新资产信息
      parameters:
      - description: 资产ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.AssetUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: 更新资产
      tags:
      - 资产管理
  /api/v1/assets/import:
    post:
      consumes:
      - application/json
      description: 批量导入资产数据
      parameters:
      - description: 导入请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.AssetImportRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.Response'
            - properties:
                data:
                  items:
                    type: string
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: 批量导入资产
      tags:
      - 资产管理
  /api/v1/assets/importance-levels:
    get:
      description: 获取所有可用的重要性等级
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.Response'
            - properties:
                data:
                  items:
                    type: string
                  type: array
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取重要性等级列表
      tags:
      - 资产管理
  /api/v1/assets/stats:
    get:
      description: 获取资产统计信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.AssetStats'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取资产统计
      tags:
      - 资产管理
  /api/v1/assets/types:
    get:
      description: 获取所有可用的资产类型
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.Response'
            - properties:
                data:
                  items:
                    type: string
                  type: array
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取资产类型列表
      tags:
      - 资产管理
...
securityDefinitions:
  ApiKeyAuth:
    description: Bearer token authentication
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"